SQLEngine: extract the SQLStandard prelude, make the engine pure
Split the ISO standard-library prelude out of the engine into a new
`SQLStandard` layer, leaving `SQLEngine` a pure engine that ships no
built-in scalar routine and seeds none. The `Routine`/`Routines` machinery,
the protected-name rule, and the dictionary-literal conformance stay in the
engine; the 14-entry `standard` prelude, its impl closures, and the shared
null-propagation helper move to `SQLStandard`.

The engine's public entry points no longer merge `Routines.standard`: `run`
and `columns` take the routines they are given (routines now required, no
implicit prelude). `SQLStandard` adds import-activated overloads —
`run(_:bindings:)`, `columns(of:validate:)` — that default the prelude, so a
call under `import SQLStandard` (or the umbrella `SQL`) resolves the built-ins
without naming them while `import SQLEngine` alone stays pure.

Two purity seams relocate to the standard layer:

  - Protected names travel WITH the value. `Routines` gains a `protected` name
    set (empty for the plain inits and the dictionary literal — the escape
    hatch a prelude is built through) that `merging(_:)` unions and
    `registering(_:…)` checks, and a `protecting(_:)` seam a prelude uses to
    mark its own names. `Routines.standard` marks its 14 names, so shadowing a
    built-in still faults SQLSTATE 42723 — now a property of the prelude rather
    than a module-wide constant, keeping the engine free of any reserved name.

  - The defined-body early-binding capture is parameterised. The engine's
    `registering(_:_:capturing:)` captures the ambient routines a caller
    supplies; `SQLStandard` adds the two-argument `registering(_:_:)` that
    passes `Routines.standard`, so a `CREATE FUNCTION` body may call a built-in
    exactly as a query would.

The umbrella `SQL` target `@_exported import`s both, restoring the
conformance-by-default a single import gave. `winmd-inspect` and the test
support import `SQLStandard`; `GUID`/`SANITIZE` stay adapter-domain UDFs in
`winmd-inspect`. The protection tests that registered a built-in over EMPTY
routines now register over `Routines.standard`, matching the relocated seam,
and a new `SQLStandardTests` suite exercises the prelude defaulting, the
protected-name guard, and the defined-body capture.
